frontend response

+++ backend/src/controllers/notes.ts
@@ -2,10 +2,15 @@ import { RequestHandler } from 'express'
 import NoteModel from '../models/note'
 import createHttpError from 'http-errors'
 import mongoose from 'mongoose'
+import { assertIsDefined } from '../util/assertIsDefined'
 
 export const getNotes: RequestHandler = async (req, res, next) => {
+  const authenticatedUserId = req.session.userId
+
   try {
-    const notes = await NoteModel.find().exec()
+    assertIsDefined(authenticatedUserId)
+
+    const notes = await NoteModel.find({ userId: authenticatedUserId }).exec()
     res.status(200).json(notes)
   } catch (error) {
     next(error)
@@ -13,8 +18,10 @@ export const getNotes: RequestHandler = async (req, res, next) => {
 }
 export const getNote: RequestHandler = async (req, res, next) => {
   const noteId = req.params.noteId
+  const authenticatedUserId = req.session.userId
 
   try {
+    assertIsDefined(authenticatedUserId)
     if (!mongoose.isValidObjectId(noteId)) {
       throw createHttpError(404, 'Invalid Note Id')
     }
@@ -25,6 +32,9 @@ export const getNote: RequestHandler = async (req, res, next) => {
       throw createHttpError(404, 'Note not found')
     }
 
+    if (!note.userId.equals(authenticatedUserId)) {
+      throw createHttpError(401, 'You cannot access this note')
+    }
     res.status(200).json(note)
   } catch (error) {
     next(error)
@@ -44,12 +54,16 @@ export const createNote: RequestHandler<
 > = async (req, res, next) => {
   const title = req.body.title
   const text = req.body.text
+  const authenticatedUserId = req.session.userId
 
   try {
+    assertIsDefined(authenticatedUserId)
+
     if (!title) {
       throw createHttpError(400, 'Note must have a title')
     }
     const newNote = await NoteModel.create({
+      userId: authenticatedUserId,
       title: title,
       text: text,
     })
@@ -76,7 +90,10 @@ export const updateNote: RequestHandler<
   const noteId = req.params.noteId
   const newTitle = req.body.title
   const newText = req.body.text
+  const authenticatedUserId = req.session.userId
   try {
+    assertIsDefined(authenticatedUserId)
+
     if (!mongoose.isValidObjectId(noteId)) {
       throw createHttpError(404, 'Invalid Note Id')
     }
@@ -86,6 +103,9 @@ export const updateNote: RequestHandler<
     const note = await NoteModel.findById(noteId).exec()
     if (!note) {
       throw createHttpError(404, 'Note not found')
+    }
+    if (!note.userId.equals(authenticatedUserId)) {
+      throw createHttpError(401, 'You cannot access this note')
     }
 
     note.title = newTitle
@@ -100,13 +120,21 @@ export const updateNote: RequestHandler<
 
 export const deleteNote: RequestHandler = async (req, res, next) => {
   const noteId = req.params.noteId
+  const authenticatedUserId = req.session.userId
+
   try {
+    assertIsDefined(authenticatedUserId)
+
     if (!mongoose.isValidObjectId(noteId)) {
       throw createHttpError(404, 'Invalid Note Id')
     }
     const note = await NoteModel.findById(noteId).exec()
     if (!note) {
       throw createHttpError(404, 'Note not found')
+    }
+
+    if (!note.userId.equals(authenticatedUserId)) {
+      throw createHttpError(401, 'You cannot access this note')
     }
     await note.deleteOne()
     res.sendStatus(204)
